-Added quadratic slider strategy implementation and tests. -Modified playback speed control to use quadratic sliders instead of linear. -Modified number formatters in player helper to use double instead of float. -Simplified slider behavior in playback parameter dialog. -Fixed potential NPE in base local fragment.

+++ app/src/main/java/org/schabi/newpipe/player/helper/PlayerHelper.java
@@ -60,11 +60,11 @@ public class PlayerHelper {
                 : stringFormatter.format("%02d:%02d", minutes, seconds).toString();
     }
 
-    public static String formatSpeed(float speed) {
+    public static String formatSpeed(double speed) {
         return speedFormatter.format(speed);
     }
 
-    public static String formatPitch(float pitch) {
+    public static String formatPitch(double pitch) {
         return pitchFormatter.format(pitch);
     }
